url shortener

+++ app/routes.py
@@ -4,6 +4,7 @@ from flask import request, render_template
 
 from app import app, db
 from app.models import ShortURL
+from app.shortener import shorten
 
 
 @app.route('/')
@@ -13,9 +14,20 @@ def home():
 
 @app.route('/add')
 def add():
-    """Create a short url via query string parameters."""
-    url = unquote(request.args['url'])
-    short = request.args['short']
+    """Create a short url"""
+    url = request.args.get('url')
+    short = request.args.get('short')
+
+    if not url:
+        return f"Error, argument <strong>url</strong> not provided", 400
+    url = unquote(url)
+
+    if ShortURL.query.filter_by(short=short).first():
+        return f"Error, <code>{short}</code> already exists", 400
+    if not short:
+        # if arg short not provided, use algorithm to shorten url
+        short = shorten(url)
+    
     obj = ShortURL(
             url=url,
             short=short
@@ -27,8 +39,10 @@ def add():
 
 @app.route('/del')
 def delete():
-    """Delete a short url via query string parameters."""
-    short = request.args['short']
+    """Delete a short url"""
+    short = request.args.get('short')
+    if not short:
+        return "Error, <strong>short</strong> argument missing", 400
     obj = ShortURL.query.filter_by(short=short).first_or_404()
     db.session.delete(obj)
     db.session.commit()
